Fixed: 1) NewFixed8 will accept as input int64 2) race condition affecting configDeafault, blockchainDefault

+++ pkg/core/util.go
@@ -170,7 +170,7 @@ func calculateUtilityAmount() util.Fixed8 {
 	for i := 0; i < len(genAmount); i++ {
 		sum += genAmount[i]
 	}
-	return util.NewFixed8(sum * decrementInterval)
+	return util.NewFixed8(int64(sum * decrementInterval))
 }
 
 // headerSliceReverse reverses the given slice of *Header.
