+++ pkg/util/fixed8_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 func TestNewFixed8(t *testing.T) {
-	values := []int{9000, 100000000, 5, 10945}
+	values := []int64{9000, 100000000, 5, 10945}
 
 	for _, val := range values {
 		assert.Equal(t, Fixed8(val*decimals), NewFixed8(val))
